pyproject.toml
[project]
name = "vega-benchmark"
version = "1.0.0"
description = "VEGA: Verifiable Enterprise GUI Agents Benchmark"
authors = [
    {name = "VEGA Team"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "playwright>=1.40.0",
    "torch>=2.0.0",
    "openai>=1.0.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "beartype>=0.16.0",
    "requests>=2.31.0",
    "tiktoken>=0.5.0",
]

[project.scripts]
vega = "vega.common.helpers:run_cli"

[tool.setuptools]
package-dir = {"" = "src"}
packages = {find = {where = ["src"]}}

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
